Exempt git management on OS X
The puppetlabs-git module does not allow specification of a
source for the git package. A source is required on OS X by
the package resource. So, the puppetlabs-git module cannot
manage git packages on OS X.

This commit sets the `manage_package` parameter of the
puppetlabs-git class to `false`. This allows ruby-build to
install on OS X, but assumes that the user has installed git
on their own.